Update: Adding password cripting fonctionnality

+++ monsupersite/App/Frontend/Modules/News/Views/show.php
@@ -20,10 +20,12 @@ foreach ($comments as $comment)
 <fieldset>
   <legend>
     Posté par <a href="/mail-<?= $comment['mail'] ?>.html"> <strong><?= htmlspecialchars($comment['auteur']) ?></strong> </a> le <?= $comment['date']->format('d/m/Y à H\hi') ?>
-    <?php if ($user->getAttribute('user')->fucType() == 1) : ?> -
-      <a href="admin/comment-update-<?= $comment['id'] ?>.html">Modifier</a> |
-      <a href="admin/comment-delete-<?= $comment['id'] ?>.html">Supprimer</a>
-    <?php endif ;?>
+    <?php if ($user->getAttribute('user') != null) : ?>
+	    <?php if ($user->getAttribute('user')->fucType() == 1) : ?> -
+	      <a href="admin/comment-update-<?= $comment['id']; ?>.html">Modifier</a> |
+	      <a href="admin/comment-delete-<?= $comment['id']; ?>.html">Supprimer</a>
+	    <?php endif ;?>
+	<?php endif ;?>
   </legend>
   <p><?= nl2br(htmlspecialchars($comment['contenu'])) ?></p>
 </fieldset>
